+++ history.rst
@@ -1,6 +1,6 @@
-####################
-Willumziehen Website
-####################
+###############################################
+Google App Engine - Address Book - Data Backend
+###############################################
 
 
 =============
@@ -10,6 +10,8 @@ Version 0.0.2
 2015-09-01
 
 - Address datamodel created
+
+- *security.ini* for authentification and authorization
 
 
 =============